metaparser-requirements.txt
dsm
python-dateutil
pyquery
